fix(*): implemented lazy loading on delete course
